requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aiohttp
aiosignal
async-timeout
attrs
cairocffi
cairosvg
cffi
charset-normalizer     
colorama
cssselect2
defusedxml
dnspython
emoji
frozenlist
discord.py
idna
isodate
lottie[pdf]
motor
multidict
natural
parsedatetime
pillow
pycparser
pymongo
python-dateutil
python-dotenv
six
tinycss2
uvloop
webencodings
yarl
